fix: linting & tests
