+++ packages/interactive-wrapper/__tests__/ios/interactive-wrapper.test.js
@@ -1,5 +1,3 @@
-import shared from "../shared";
 import nativeTest from "../interactive-wrapper.native";
 
-shared();
 nativeTest();
