+++ packages/interactive-wrapper/__tests__/web/interactive-wrapper.test.js
@@ -1,3 +1,40 @@
-import shared from "../shared";
+/* eslint-env browser */
+import React from 'react';
+import {
+  addSerializers,
+  compose,
+  flattenStyleTransform,
+  minimalWebTransform,
+  enzymeTreeSerializer,
+  print
+} from "@times-components/jest-serializer";
+import { mount } from 'enzyme';
+import InteractiveWrapper from "../../src/interactive-wrapper";
 
-shared();
+addSerializers(
+  expect,
+  compose(
+    print,
+    flattenStyleTransform,
+    minimalWebTransform,
+  ),
+  enzymeTreeSerializer()
+);
+
+it("renders correctly", () => {
+  const props = {
+    attributes: {
+      chaptercounter: "Chapter%20one",
+      heading: "A heading",
+      standfirst: "A standfirst"
+    },
+    element: "chapter-header",
+    id: "a0534eee-682e-4955-8e1e-84b428ef1e79",
+    source:
+      "//components.timesdev.tools/lib2/times-chapter-header-1.0.0/chapter-header.html"
+  };
+  const container = document.createElement('div');
+
+  mount(<InteractiveWrapper {...props} />, { attachTo: container });
+  expect(container.innerHTML).toMatchSnapshot();
+});
